Create password goes to `#create` method
**Why**: We heart rails
