+++ app/controllers/sign_up/confirmations_controller.rb
@@ -1,23 +1,10 @@
 module SignUp
   class ConfirmationsController < Devise::ConfirmationsController
     include ValidEmailParameter
+    include UnconfirmedUserConcern
 
     def new
       @user = User.new
-    end
-
-    def confirm
-      with_unconfirmed_user do
-        result = @password_form.submit(permitted_params)
-
-        analytics.track_event(Analytics::PASSWORD_CREATION, result)
-
-        if result[:success]
-          process_successful_password_creation
-        else
-          process_unsuccessful_password_creation
-        end
-      end
     end
 
     def show
@@ -35,27 +22,6 @@ module SignUp
     end
 
     protected
-
-    def with_unconfirmed_user
-      token = params[:confirmation_token]
-
-      @user = User.find_or_initialize_with_error_by(:confirmation_token, token)
-      @user = User.confirm_by_token(token) if @user.confirmed?
-      @password_form = PasswordForm.new(@user)
-
-      yield
-    end
-
-    def process_successful_password_creation
-      @user.confirm
-      @user.update(reset_requested_at: nil, password: permitted_params[:password])
-      sign_in_and_redirect_user
-    end
-
-    def process_unsuccessful_password_creation
-      @confirmation_token = params[:confirmation_token]
-      render :show
-    end
 
     def process_successful_confirmation
       if !@user.confirmed?
@@ -99,27 +65,6 @@ module SignUp
         @user.decorate.confirmation_period_expired_error
       else
         t('errors.messages.confirmation_invalid_token')
-      end
-    end
-
-    private
-
-    def permitted_params
-      params.require(:password_form).permit(:confirmation_token, :password)
-    end
-
-    def sign_in_and_redirect_user
-      sign_in @user
-      redirect_to after_confirmation_path_for(@user)
-    end
-
-    def after_confirmation_path_for(user)
-      if !user_signed_in?
-        new_user_session_url
-      elsif user.two_factor_enabled?
-        profile_path
-      else
-        phone_setup_url
       end
     end
   end
