+++ app/views/sign_up/confirmations/show.html.slim
@@ -10,7 +10,7 @@ p.mb0
   = t('instructions.password.info.point_2')
 = simple_form_for(@password_form,
     url: sign_up_create_password_path,
-    method: :patch,
+    method: :post,
     html: { role: 'form', autocomplete: 'off' }) do |f|
   = f.input :password, required: true,
       input_html: { 'aria-describedby': 'password-description' }
