+++ spec/support/features/session_helper.rb
@@ -80,7 +80,7 @@ module Features
         confirmation_token: @raw_confirmation_token, confirmation_sent_at: Time.current
       )
       visit "/sign_up/create_password?confirmation_token=#{@raw_confirmation_token}"
-      # visit user_confirmation_url(confirmation_token: @raw_confirmation_token)
+      # visit user_confirmation_path(confirmation_token: @raw_confirmation_token)
     end
 
     def sign_up_and_2fa
